[Issue-3864] Refactor identical code
