+++ packages/extension-koni-ui/src/components/AccountProxy/index.ts
@@ -10,4 +10,5 @@ export { default as AccountProxyAvatarGroup } from './AccountProxyAvatarGroup';
 export { default as AccountChainAddressItem } from './AccountChainAddressItem';
 export { default as AddressSelectorItem } from './AddressSelectorItem';
 export { default as AccountProxyTypeTag } from './AccountProxyTypeTag';
-export { default as AccountChainAddressWithStatusItem } from '../AccountProxy/AccountChainAddressWithStatusItem';
+export { default as AccountChainAddressWithStatusItem } from './AccountChainAddressWithStatusItem';
+export { default as AccountChainAddressList } from './list/AccountChainAddressList';
